fix: fix review points
